modify: /api/user/favorite route 수정
관심 종목을 해제하는 로직을 post로 이관하고 DELETE /api/user/favorite route 제거

+++ back/src/api/user/favorite.ts
@@ -20,21 +20,15 @@ export default (): express.Router => {
 	router.post('/favorite', sessionValidator, async (req: Request, res: Response, next: NextFunction) => {
 		try {
 			const { userId } = res.locals;
-			const { stockCode } = req.body;
+			const { stockCode, shouldDeleteFavorite } = req.body;
+
+			if (shouldDeleteFavorite) {
+				await UserFavoriteService.removeUserFavorite(userId, stockCode);
+				res.status(201).json({});
+				return;
+			}
+
 			await UserFavoriteService.createUserFavorite(userId, stockCode);
-
-			res.status(201).json({});
-		} catch (error) {
-			next(error);
-		}
-	});
-
-	router.delete('/favorite', sessionValidator, async (req: Request, res: Response, next: NextFunction) => {
-		try {
-			const { userId } = res.locals;
-			const { stockCode } = req.body;
-			await UserFavoriteService.removeUserFavorite(userId, stockCode);
-
 			res.status(201).json({});
 		} catch (error) {
 			next(error);
